fix(etl): enhance etl to process larger batch for higher efficiency
